fixed schooldirectory table
